Rename automaton package to syntax_check
