+++ pgpainless-core/src/test/java/org/pgpainless/decryption_verification/syntax_check/PDATest.java
@@ -1,4 +1,4 @@
-package org.pgpainless.decryption_verification.automaton;
+package org.pgpainless.decryption_verification.syntax_check;
 
 import org.junit.jupiter.api.Test;
 import org.pgpainless.exception.MalformedOpenPgpMessageException;
